requirements.txt
Flask
python-dotenv
watson-developer-cloud